adapt freeze v2 case
